+++ lib/models/geolocation_model.dart
