+++ lib/models/path_overlay_model.dart
